+++ Changelog.rst
@@ -12,6 +12,7 @@ Changelog
  * added support for proper SSL certificate handling
  * Required parameters are now checked for non-empty values
  * ConnectionPool now checks if any connections were defined
+ * fixed a race condition in ConnectionPool
 
 1.2.0 (2014-08-03)
 ------------------